webapp/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
marshmallow==3.20.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
